gui_e2e: fix failing email notifications test
The fix consists of moving away from accessing elements by index to
instead use the aria labels when they exist. We also want to configure
the new entry which is created instead of the default entry. If we don't
do this, the rule never gets created because the added entry is left
dangling and unconfigured.

CMK-21678

Change-Id: I183e9b7285cdbfaef97362770238722840d50eee