check-http: show user friendly message for --server errors
When parsing arguments, we catch these errors and display them
in the check result.  However, the messages in some case are
not useful to the enduser using the UI. This change will modify
the error message shown for --server errors.

So instead of showing the following
error: invalid value '10.201.2.258' for '--server <SERVER>': Invalid server address: 10.201.2.258 (!!)

We now show
Invalid server address: 10.201.2.258 (!!)

CMK-21983

Change-Id: I36045c48f74ba21be732b07aee330ad9274cf64d