17495 SEC Stop python-ldap from logging clear passwords
We are currently enabling tracing and OPT_DEBUG_LEVEL in python-ldap /
the underlying libldap library.

This will log all internal method calls and their arguments and
send them to stderr. This in turn will show up in our apache error_log.

As there is no fool-proof way to only hide sensitive information with
the current tools, we completely remove this functionality.

For reference, here's short discussion on this topic:
https://github.com/python-ldap/python-ldap/issues/384

SUP-22379

Change-Id: I3c5f9eba7cf92526964cc12d1fbdc406603727b2