Automation helper: Separate result of automation call from stdout
Before, the automation helper received the serialized result via stdout. This is
unnecessarily complicated and also runs the risk of mixing up log messages and
the serialized result.

Now, the automation helper receives the result via a function call. Note that we
still capture stdout and forward it, st. it can be reported, which might
faciliate error diagnoses.

CMK-21736

Change-Id: I5fcd740503324225f84ecfcc4924de2d21106b19