gui_e2e: make email notifications test more robust
Instead of asserting that the count is zero or not zero, let's take into
account the current stats before running the test. Now we are asserting
that the total notifications sent count has changed and the total
failures has remained the same.

This makes the test more robust because now notification failures from
other tests don't have an influence on what we are asserting in this
test.

CMK-21533

Change-Id: I79728f2929d83421518fadd4b4280f312e442367