Mount complete cache folder of HOME as tmpfs with exec option
This saves another 1GB of diskspace and shrinks the container_shadow_workspace
to only a few hundred MB at the price of recreating everything from scratch
whenever a container is entered. With CMK-21804 a container is only entered
once for a complete Gerrit CV chain run.

This follows CMK-21123

CMK-21532

Change-Id: Ia3f7ebd69c804f6c31dc6359a7c8718d7fb24cd0