Add option to test connection to AWS agent
Usually exceptions that are relevant for the user (such as incorrect credentials) are passed to the "AWS Exceptions" service. The Quick Setup however expects an error exit code for the agent to notify the user early on that the connection could not be established.
Add a dedicated option for the Quick setup to use to the agent, similar to the Azure agent.

CMK-21198

Change-Id: I05f950c72175d5e02543dfa92870d25a268d17a1