17620 FIX agent_aws: Do not crash when invalid credentials are given
If e.g. a wrong secret access key was provided to the AWS agent, it would result in a crash with the error message
```
ClientError (An error occurred (SignatureDoesNotMatch) when calling the GetCallerIdentity operation: The request signature we calculated does not match the signature you provided. Check your AWS Secret Access Key and signing method. Consult the service documentation.)
```

Errors due to invalid credentials are now directed to the "AWS Exceptions" service instead as with other errors that occur when accessing AWS and require user attention.

CMK-21198

Change-Id: Iaf5d12b4ec73810dbdbd33f11128d388ff1ff326