17587 FIX informix_dbspaces: fix calculation of free and used space size
Prior to this werk, the free and used space size was incorrectly calculated as the chunk pagesize was taken
instead of the dbspace default pagesize. This werk fixes this issue by querying the correct pagesize value. To
benefit from this change, the mk_informix plugin should be reinstalled with this new change. The check plugin
itself remains unmodified.

SUP-22249

Change-Id: Iece8cff4ba2b014185c492fd98fd845faa22bdb7