Improve piggyback processing performance
A call to get_piggybacked_hosts_time_settings with a given piggybacked_hostname
processed all piggyback data needlessly, leading to a huge performance
impact.

This change propagates the piggybacked_hostname further to reduce the
overhead in the called helpers.

We'll have to see if this is enough. If not, we will need to make more
drastic changes ;).

CMK-21777

Change-Id: I78d72f13518c65acdba72b99cab0c7ec38d3cb23