Background jobs: Prevent race condition during startup
The start was roughly executing the following steps:

1. Check if job is alive (by checking work directory and asking the job scheduler)
2. Initialize the job work directory
3. Call job scheduler to start the job
4. Mark the job as running in the job scheduler

This is not race condition free. Independent job starts could interfere
with each other by e.g. removing the already initialized work directory
of the other job run.

This change now moves the whole initialization into the job scheduler
which protects the initialization steps with a lock. For this, the
scheduler needs to do a bit more than just managing the jobs thread. It
additionally needs to initialize the working directory.

CMK-21775

Change-Id: Ia51f3669cb2f1aeaa75354b096b32a5bde070747