Fix migration of rules including passwords
It was possible that `cmk-update-config` failed because of an
exception in `cmk --automation update-passwords-merged-file`.

This exception was raised, if multiple old rules (which have to
contain passwords) in multiple folders belonged to a migrated
rulesets. After saving every `rules.mk` file, the automation
command was called.

So, after saving migrated rules in the first file the hook
command would be called. Processing the non-migrated rules in
the other files triggered the command into raising an AttributeError.

This is now fixed by saving all `rules.mk` files first and
calling the hook command only once afterward.

CMK-21820

Change-Id: I3175473b2d8b73578de091e514bf675b05234f17