migrate_notifications: improve test coverage
Currently, we have unit tests for the respective plugins, i.e. msteams.
However, we also want to test the edge cases that were causing failures
during manual tests. Namely:

  - already migrated notifications (uuid or None).
  - partially migrated notifications.
  - custom plugins that have a list as parameter argument.

These are are currently marked as xfail, but we'll want to fix these in
an upcoming change, especially the first two, as they are the ones
creating failures.

CMK-21849

Change-Id: I2630b969f2cc8aed1e72471e7785a739b5236191