quick_setup: Catch 429 error during save action
When a 429 status code is received, the error handler will return a
CmkSimpleError containing the following message:

Cannot start action: Another Quick setup action already running.

This error message is formed with the title and description of the error
received from the backend and is translated to the user's language.

CMK-21767

Change-Id: I72f2dd205e0354cae877b72b659ea0e519b2a9d2