Background jobs: Use pid field again to store the jobs thread ID
The pid field was not populated anymore since we moved the jobs to
threads. With this change, the pid field is used again to store the
thread ID issued by the kernel, which may help debugging in the future.
The ppid field is now also set again, typically holding the process ID
of the UI job scheduler.

Change-Id: Ibe1f10ad942f02387646740c5ff211d3ceae6c2e