17006 FIX View filtering: Preserve necessary filters on "Reset"
Clicking the "Reset" button within a view's filter popup used to clear all filters which were not given in the builtin view specification. This led to filters necessary for rendering to be reset as well.
E.g. in the view "Services of host _hostname_" the filter "Host name (exact match)" was cleared upon reset, leading to an error message.

This behavior is fixed to preserving filters which are necessary for rendering a given view when clicking on "Reset".

Change-Id: I440a7379348afed42a4454efc4c2456c62ebdbb1