17665 check_http: Postponed hard deprecation of old HTTP monitoring plug-in
In werk #15515 the deprecation of the old HTTP monitoring plug-in has been
announced for Checkmk 2.4.0 release. For several reasons, the deprecation
is postponed to 2.5.0 release. This gives more time to migrate to the new
plug-in rule set.

All other statements remain (now, beginning with Checkmk 2.5.0):

- The plug-in file is still shipped with Checkmk and accessible with the
  generic rule set for active checks.
- No new rules can be created in the deprecated rule set; all existing
  ones are still used

This werk is considered as incompatible as the migration needs to happen
before 2.5.0 release as a migration cannot be rolled back afterwards.

Change-Id: Ib4a2807d0185f48c65eea9efab780f1848a83860