Keep config cache as soon as it is created.
This removes most calls to `get_config_cache`.

The result is still somewhat backwards (see
comment in LoadedConfigFragment), but helps
us to distinguish the callsites where we call
it to actually create a ConfigCache instance
from the ones where we merely access the current
one.

Note that if we had to pass arguments to the
ConfigCache creation, these functions would have
different signatures.

Change-Id: I82f141870d882b3853673f676feab28f213ef050