17047 FIX systemd_units: Properly exclude ignored services from failed count in systemd summary
In systemd summary checks, services that were configured to be ignored in the ruleset were still affecting the state of the service. This was not the expected behaviour, as ignored services should not affect the state of the service.

Now, if a service is configured to be ignored in the rule:
* It will only appear in the "Ignored" services count

This is a follow up to werk #17042, where the unintended behaviour was introduced.

Change-Id: I4b4c893abaf33d48de2c0c79f6ee3f4a9d0bdd0e