Remove autoflake, pyflakes and flake8 leftovers
The most important features of those tools are now covered by ruff,
which is a first class citizen and properly integrated into the CI and
development workflows.

Flake8, autoflake were not properly integrated into the CI, so they
appeared to be used only locally and only by a few developers. So there
seems to be not too much value left in keeping them around.

Change-Id: Ifed39145b42cc989541388a0ffa11eb3b5782699