gerrit: rename section collector
For extensibility purposes, this collector should only concern itself
with fetching the data necessary for the gerrit version section. This
gives offer flexibility when adding additional collectors.

In addition, we can specify how we want to cache each collector. For
example, the gerrit version changes much less frequently than commits to
a repository.

Change-Id: I5ed2b81bfbd71a2e502538c1a667782a60f61b84