Addendum for werk 17676: Revert handling of stacked elements
Werk 17676 changed the parsing of bidirectional graphs. Elements from the upper
graph are now added before elements from the lower graph. The intentation was to
fix the order of elements in the graph legend.

However, there was one caveat: Before this commit, for any graph containing at
least one stacked element, the legend order was reversed. Hence, werk 17676
broke the (previously correct) order of graph legends with only stacked
elements.

This commit restores the order in which stacked elements are added back to the
way it was before werk 17676. However, note that this is only correct for some
cases (such as only stacked elements). In 2.4 and master, we implement a full
fix, which is however deemed too risky to fully backport.

CMK-22004

Change-Id: I2befd50ab633b38ab9f6e8adcaa007b3e60bb71c