17682 FIX Fix discovery crash when mssql_jobs don`t have a name
Closes #796
CMK-22265

Some jobs do not have a name so that field is missing in the agent plugin
output (both, the deprecated mssql.vbs and the new mk-sql.exe) and will
be set to an empty string in the parsed output. That leads to a discovery
crash and all subsequent jobs will not be discovered as well. As a hotfix,
a check whether job is unnamed was added to prevent the discovery process
from crashing.

Co-authored-by: Diana Ivanchevska <diana.ivanchevska@checkmk.com>

Change-Id: Ib2fe20ee8f391b49f045850c43f5cbec49909d52