remove ErrorBoundary
looks like a relict from the early days when EditReadonly was unstable.

the problem is, that useErrorBoundary will catch all errors of this
component and all child components used. But the catched errors will
only be shown to the user when they finished the stage and show the
recap.

CMK-21752

Change-Id: Ifcc3eff91746c06702dfe864359148f37848f3d3