remove pointer-events: none from QuickSetupStage
the reason for this change is, that an error was shown in the recap, but
the details-button was not clickable because of the `pointer-events: none`.

the idea of the old variant was to make the :before pseudo element
clickable. but this technique already bit rotted, so we decided that the
title is enough and thus could simplify the styles a bit.

CMK-21752

Change-Id: Ic5a93fb456a348cb5dc069e8a4ea133129ee6c51