18272 Extend interface discovery to allow discovery of multiple service labels
This changes the ruleset _Network interface and switch port discovery_ and the Discovery itself to allow multiple service labels
with individual matching conditions to be defined in a single rule.

Prior to this Werk the ruleset only allowed a list of service labels that would be discovered if the
"Conditions for this rule to apply" would apply.
If you would want to discover multiple service labels based on different conditions, you would need multiple
rules and with any one additional service label the amount of rules would be increasing exponentially.

This Werk introduces an additional setting that extends the service label discovery by allowing
multiple sets of service labels with explicit matching conditions (index, alias and description).

Thanks lbuhleie for this contribution!

closes #798

Change-Id: Idebc5ca3ecd4184e8acb7b24bc1d5df22b4d4d0b