15356 FIX agent solaris: implement alterantive way to get total available memory
This is a followup to [Werk #15348](https://checkmk.com/werk/15348) which made
the statgrab sections optional.

The statgrab output was used to determine the total amount of ram being
available on the system. This is important for the ps check where you can
configure thresholds in percent of total ram available.

As the statgrab output is optional, the ps check was affected when the
statgrab output was not available and showed "Percentual RAM levels configured,
but total RAM is unknown"

Now the total amount of RAM is received via `prtconf` and thus independent of
the statgrab section.

Change-Id: Ifc84b8c3092d77ab20a0cde5bfa2e055c3e29693