rulespecs: Add possibility to announce a future deprecation
Until now, we only have the possibility to actually deprecate
a rulespec. This has the effect that the user can't create a new
rule of that rulespec anymore.

With the new flag, it's possible to announce a future deprecation.
This has the effect that a clearly visible warning will warn the
user to not use the rulespec anymore, while generally allowing
the usage.

CMK-21174
CMK-21175

Change-Id: I4fc4aa0323d9a07c239c9951b82bd3c4143a7ce9