17720 solaris: fallback to uname system binary
Solaris ships two versions of the `uname` program:

  - Solaris specific program (/usr/bin/uname)
  - GNU program (/usr/gnu/bin/uname)

With the prior implementation, we were relying on Solaris specific
`uname` to be properly resolved via the `$PATH` variable. Now, we're
first testing to see if it is available. If not, we invoke uname with
a call to its full path: `/usr/bin/uname`.

In the event that the underlying `/usr/bin/uname` has been altered, we
will not write out anything to the agent output as opposed to failing
during check parsing.

SUP-22439

Change-Id: I2b62f7873214980fda67557f49fc2ad4cce41b88