Removed dead code for handling active check results
The modes "check", "check-discovery", and "inventory-as-check" had a handler
for the active check results.  For the CRE, the handler was a no-op, and for
all other editions it added the result to a global variable in the keepalive
code.  But those modes do not use the keepalive code, so the modifications
of the global variable were never observed.  As a consequence, the handler
could *always* be a no-op, in other words: It could be removed completely.

This simplifies the mode-related code quite a bit, removing some edition
hacks.  In addition, all use sites of the global variable are now completely
within the keepalive code, so we can further simplify things there,
hopefully removing the global variable hack completely.

Change-Id: Id922b10df1e7b5051281c9e9dd4cc8e05dea304a