17601 Add initial info text on "must search" views
SUP-21395

For "must search" views, an initial filter search must be performed to display the contents of the view.
Currently the view area is empty and there is no indication of this behaviour.

An info text has now been added to indicate that the filter search must be performed first.

Change-Id: I3e5c5a943a11f66b45cc51fcab913e7e70dc6afd