pre-commit: use ruff version defined in bazel
By using the community ruff hook in pre-commit, we are susceptible to
version drift. For example, right now the version of ruff that is used
by bazel and CI is v0.9.6, while the pre-commit has lagged behind at
v0.8.3.

By using `scripts/run-uvenv` as our entrypoint, we make bazel the
single-source of truth for ruff.

Change-Id: I074b0d2ca0ffcda682c4475ab054fc781355dcea