Cron job scheduling: Avoid accidentally skipping jobs
The cron job scheduler is implemented to ensure a fixed delay between two
subsequent scheduling cycle starts. However, there are no guarantees on the
delay between two subsequent checks if a job is due.

During each cycle, to decide if a job is due, the scheduler checks if the last
start time of this job is longer ago than its interval. Due to the behavior
described above, the previous scheduling cycle interval of 60 seconds lead to
to jobs with a scheduling interval of also 60 seconds being skipped by accident.

As a quick fix, we decrease the scheduling cycle interval to 5 seconds.

CMK-22197

Change-Id: I18f852122ebac75153644824f50bf33a8c7fd1e9