17580 activate_changes: add activation status to rest-api response
We have introduced a new field 'status_per_site' to the response
in two of our activation_run endpoints. This new field will hold
a sequence of status objects, one for each site. In this response
we will include the activation status, details, the site the
activation was peformed on, the start time and the end time.

The endpoints affected are:

/domain-types/activation_run/collections/running
/objects/activation_run/{activation_id}

CMK-21754

Change-Id: I451f846b7e5d44b7ab9b311bffe4fdb4a469e30d