docs: remove call to pre-commit for linting
Firstly, it is inconsistent to offer this option for the linting command
and not the formatting command. Secondly, the call to pre-commit should
not be recommended used when a better alternative is available.

Pre-commit should really only be invoked in the context of a commit. By
calling it explicitly, you would need to pass files explicitly. So the
command that is currently being shown is essentially a no-op anyway.

Change-Id: I761cd97f96cb8537efb4e5b87288a8bbc776f082