17388 FIX mk_logwatch: Activate bakery plugin for AIX
In contrast to the title _Text logfiles (Windows, Linux, Solaris, AIX)_, the _mk_logwatch_ bakery plugin previously didn't ship the agent plugin with AIX agent packages.
The reason was that we didn't thoroughly test the plugin for AIX, so the reference in the title was added erroneously.

Since we have some feedback from Checkmk users about the _mk_logwatch_ plugin working as expected on AIX, we now activate the bakery rule for AIX.
Please note that you need a recent Python 3 installation on the target AIX system in order to operate the plugin properly.

SUP-21309

Change-Id: I63cfabf27f67e9a96976b852b7bd65a46555dcd6