17528 FIX tcp_connections: Support for connection data from `ss` and `netstat`
The `ss` command has a different order of columns in the output than `netstat`. This caused a ValueError when creating a monitoring rule for UDP or TCP connections, when the connection data is provided by the `ss` command.

This fix parses data from the `ss` command correctly to avoid the exception.

SUP-22678
Change-Id: I5fda1eeb25112e4f0507d52f31b6233c4fe2ce50