Get rid of selectedSuggestionIndex in CmkSuggestions
also:
* use emits instead of callbacks
* use @update:model-value instead of watcher
* remove unnecessary <template>
* be more specific in unittest (not sure how this worked before)

The previous implementation filtered the elements based on the index of
the elements. This meant you had to handle two structures: the actual
suggestions and then a list of pointers (indices) to those suggestions.
In certain situations those two data structures were not kept in sync and
some bugs surfaced.

This change now removes the need to have two independent data structure
with the penalty that there is a slight runtime overhead when using
the keyboard navigation.

CMK-21802

Change-Id: Id959df491bd9df61f4580608f812a3d280fe395b