17710 introduce possibility to update discovered service parameters
Prior to this werk, it was not possible to update only the discovery parameters
of a service. Updating those were tied to updating the identified parameters
of the service.

This werk introduces the option to update the discovery parameters.

The following changes were made to introduce this option as well for better
visibility:

* Service discovery page: changed services due to a change of their discovery
parameters are now listed under 'Changed services'.
* Specific service discovery actions can be used to update the parameters of all
changed services or target update a specific service.
* `Check_MK Discovery service`: the service now accounts for changed services
due discovery parameter changes. The service state of the discovery
service for such cases can also be configured.
* Bulk discovery & Periodic service discovery: now have the option to update
discovery parameters of changed services

Previously, actions on the service discovery page affecting monitored services
did not apply to changed services; however, since changed services are also
considered monitored services, they should have been affected.
With this Werk, this inconsistency was fixed.

CMK-22177

Change-Id: Ibe2b93a5c646257c519571813655214267d0ecb6