16723 FIX remove invalid host labels on site update
In Werk #16719 a validation was introduced in the REST API that
prevents entering host labels with `:` in the name or value (e.g.:
{‘ke:y’: ‘value’, ‘createdAt’: ‘12:00:00’}.
However, it was not taken into account that there may be invalid labels
entered in a lawful way before this Werk, leaving it up to the user to
remove such labels.

This Werk introduces an upgrade hook that removes invalid labels
automatically when the site is upgraded to a new version.

SUP-22482

Change-Id: I0e434d5c5881d7005cca30a40dfbfee44de39097