17723 FIX Printer Supply perfometer color
The perfometer relies on the service description to determine its color:
black, yellow, cyan, magenta, or grey (unknown). This works as expected
when the language of the description is in English, but fails if the
color is in a different language.

This is a small hack to get this working in 2.3. However, this
cannot be applied to 2.4 and above as the architecture has completely
changed.

SUP-22762

Change-Id: Ib609980af2d8d98480a562f76c9d3fd2327d2f85