openapi: introduce endpoint_family_registry
In the current API structure, all endpoints must be defined in the same
module as the spec generator infers the necessary title and doc body
based on the docstring. This has some disadvantages namely definition
flexibility being one of them. This commit introduces the (for now)
optional endpoint family definition to overcome this restriction.

Change-Id: Ia5378ab2ee63ebf500c628e3dcac4177a4d9a96f