17747 FIX plugins: Show unknown state if a kubernetes cronjob has a invalid/unknown state instead of raising an error
When using the kubenertes cronjob check it was possible in rare cases to create a situation that was invalid/unknown to the plugin, in which case a unhandled error was raised.

Instead of an error we return a Unknown State for the check now in cases where we encounter this situation.

Change-Id: I8a13be4658f072edfa3fdcb81f36fd58994fb312