17581 FIX View: Events of host and services log details
This werk addresses an issue when displaying log details
within the Events of host & services view.

Previously it was incorrectly showing an error when no
html text was part of the log details.

"HTML output can not be rendered because of truncated
data. (Increase limit for future entries)WARN"

With this werk, this error is no longer shown.

SUP-21088

Change-Id: I9e76266b3b46ca459fe56e6430d658d5a4dfa339