17008 FIX View rendering: Prevent scrollbars from blocking view content
Views, e.g. "All hosts", are rendered with a scrollbar in case their content exceeds the given browser space. Scrollbars are rendered overlaying the content, which could result in some invisible and/or unclickable view content.

A views content is now given enough padding to the right and the bottom so the scrollbars do not block content anymore once users scroll the given content to the end.

SUP-22429 SUP-22395

Change-Id: I5f098ac049784980c616c61e09c57741fb705ac9