17603 FIX Check logged_out on single user session mode on session init
SUP-22315

When single user session mode was enabled, it was not possible to log in again before the session timeout expired.
This behavior was caused by ignoring the status "logged_out"=True in the session_info.mk file.

This status is now checked correctly, allowing immediate re-login without waiting for the session timeout.

Change-Id: I446da5b5ca93a053946ae7507527d9e15f08d2e8