diskstat discovery: keep item
CMK-17090 / Werk #16857 caused a change in the naming of services. This change restricts the change to diskstat plugin.

**What is intended with Werk #16857.**

If there is an existing configuration it will get the legacy option `Use device name as service name`. In the words of the werk:
```
Old `Physical disks` configurations will continue to use device name as service description to preserve compatibility. For the new configurations using WWN will be the default.
```
Thus, no change in behaviour should be observable, if one simply updates the agent (or updates to 2.4.0).

**What is definitely broken**

Existing checks, which use `discovery_diskstat_generic`, which have colons in their device name.

**What this change does**

Currently, no check plugin except `diskstat`, benefits from Werk #16857. This is because they don't implement the section block `[device_wwn]`. With this change the behavior of these plugins is restored.

**What this change does not do**

The `diskstat` plugin behaviour was not touched by this change. It is not clear to me how
* <<<multipath>>>
* DISKSTAT_DISKLESS_PATTERN
* lvm
* vxvm
are affected to Werk #16857, but I was told it works as intended TM.

CMK-21985

Change-Id: I25c5b25bfcf2652eec91e36494ace4906595bcee