17582 FIX downtimes: users with access to services only, should not be able to set a downtime on the host
Previously, when a user assigned to a contact group that was in turn
assigned to services via a rule, they could create a downtime on the
host the services belonged to but then didn't have permission to
remove the downtime.  This werk addresses this issue by not allowing
the user to set the downtime if they do not have persmission to do so.

SUP-22380

Change-Id: Ibca3885a9f0d5bcc96446e4f1e0ab90fbadd4782