Move HW/SW inventory logic to active check plugin
The active check plugin `lib/nagios/plugins/check_cmk_inv` was
introduced in 96fb83f5 as wrapper for `cmk --inventory-as-check`. The
change introduced some overhead because it now requires starting two
processes and two python interpreters.

This change now moves the implementation of the `--inventory-as-check`
mode to the active check plugin and removes the argument from the `cmk`
command. Even if we change the command line of the `cmk` command, we
don't mark this as incompatible change, because we consider the argument
as an implementation detail.

CMK-22194

Change-Id: Ia1852a05622cd4491631ac6ad9a5209480ceb57c