Run user profile cleanup as threaded cron job
Before, the cleanup job ran non-threaded, but launched a (threaded) background
job. This is unnecessary and comes with a potential race condition: When the UI
job scheduler starts, it first spawns off the cron job scheduler. Then, it
proceeds with starting up the API service for launching background jobs. Hence,
there is a small time window where cron jobs might already be executed but the
API for launching background jobs is not yet reachable.

Change-Id: I3711c54aad74f233c1de97a5c8c923404a2a310f