Explicitly do TLS shutdown when closing socket
Apparently the socket does not perform a proper TLS handshake on
teardown. Our stunnel therefore complains about that with

    SSL_read: ssl/record/rec_layer_s3.c:322: error:0A000 126:SSL routines::unexpected eof while reading

Apparently stunnel waits for a `close_notify` but that never comes. It
seem that this `close_notify` was more a nice to have but openssl starts
to get stricte (which is a bit pointless on closing a connection IMHO).

Anyways if we unwrap the socket we don't cause these logs.

CMK-21171

Change-Id: Ib5830a5ca1da0e6ebd288a2a0d020a244ad58f91