remove `OMD_TO_VERSION`, `OMD_FROM_VERSION` and `OMD_CONFLICT_MODE`
`OMD_CONFLICT_MODE` is passed via the command line, so we don't need an
additional way.

The `OMD_TO_VERSION` and `OMD_FROM_VERSION` require the function
`_omd_to_check_mk_version` to be parsed. This function is internal to
OMD, and thus the variables should not be exposed to
`cmk-update-config`.

Moreover, `OMD_TO_VERSION` should never be used: `OMD_TO_VERSION` always
matches the "version the code". If code shouldn't be executed in a
certain version, then that code should be statically excluded from that
version of Checkmk.

CMK-22299

Change-Id: Ib909a3792a1b1f22997fbfeea873962920337d79