16572 FIX Replace cmk --inventory-as-check with check_cmk_inv command
The command `cmk --inventory-as-check`, that can be called on the command line
by the Checkmk site user, has been moved to a dedicated check plugin entry point
`lib/nagios/plugins/check_cmk_inv`.

The check plugin entry point already existed in 2.3, but was so far only a thin
wrapper around the `cmk --inventory-as-check` command. Now the command has been
removed and the check plugin entry point is the one that is used by the
monitoring core to execute the "Check_MK HW/SW Inventory" service. If you want
to call this logic manually, e.g. for debugging, you need to use the new entry
point now.

Even if we change the command line of the `cmk` command, we don't mark this as
incompatible change, because we consider this as an implementation detail.

Change-Id: Id4ac3b949a21f201d8beeae39b55f6a8d8291c49