17813 Remove transaction id & csrf token from sorting url
SUP-19542

The sort links of tables and views were generated with a transaction id and a csrf token.
This is not required for the pure sorting of a table/view and sometimes led to unforeseen redirects when sorting tables or views.

These GET parameters have now been removed from all sort links.

Change-Id: I802eb99af59a7dd1db9f064d7e3385dea712c5a3