quick_setup: improve type for icon name
By using the type `str` as the name type, we don't inform the caller
which icon names are valid. By exposing the `IconName` type, we can
assist callers by constraining input to only valid icon names.

The `IconName` type values are derived from the following css file:

  `packages/cmk-frontend-vue/src/assets/variables.css`

There is currently no way to keep these files in-sync. For now, we'll
need to manually update the type. But in the future, we may want to add
this to the shared typing library.

Note: no runtime validation is being done for this value, this is only
to assist developers to use a correct icon name. Only mypy or their
respective LSP will complain when an invalid value is passed.

Change-Id: Ideb02238087324d9be7e3d2fe31ec9e73d5618b8