openapi_session: Adapt CMK version comparison
Some API calls differ between CMK v2.3.0p* and CMK v2.4.0*. This is the
reason why we need to include some version-comparison logic within
helper methods in this module.

The CMK version-comparison logic should also work with the newly
introduced 2.4.0b* versions.

Being beta-versions, we have that 2.4.0b* < 2.4.0.
In this change we fix this problem.

Change-Id: I6cc4a8f17fcb26510fa3db330e0e2846d78cbbb8