17825 FIX Prevent availability view crash due to division by 0
With long time periods (e.g. last year) and recently added hosts, the availability view crashed when the "Include unmonitored time" option was disabled.

This was due to the fact that the monitored time of the host tended to be 0 in this constellation.

In this case the specified time instead of the span is now used to calculate the uptime in percent.

Change-Id: I687629d1f198669a4a92c5c1eea1b5e720290a8f