17816 FIX downtimes: users with access to host services but not the host, should not be able to set a host downtime
Previously, when a user assigned to a contact group that was in turn
assigned to services via a rule, they could create a downtime on the
host the services belonged to but then didn't have permission to
remove the downtime.  This werk addresses this issue by not allowing
the user to set the downtime if they do not have persmission to do so.

SUP-22380

Change-Id: Idef309e9870caa011a1ee58ac8d1442f75a3d895