FIX 17786 Network Interface and switch port discovery ruleset: MICROSENSE G6 and Cisco Cataclyst 9200L switches alias naming
This fix picks up aliases in SNMP discovery for MICROSENS G6 switches
and Cisco Catalyst 9200L switches and allows to use these aliases for
naming in the discovery rule.

When "alias" was used in "Network Interface and switch port discovery"
for "Appearance of network interface", but the device was discovered internally
as a device with only the "ifTable (.1.3.6.1.2.1.2.2.1.*)" tree available, the
alias was not picked up as part of the "ifAlias (.1.3.6.1.2.1.31.1.1.1.18)" tree.

CheckMK normally only picks up those values when other 64-bit counters are available.
Some device manufacturers still provide aliases despite not providing 64-bit counters.

CMK-12330
SUP-22935

Change-Id: Ic2cc94011e0f21f0685495aa999f60a49764b98a