Do not synchronize `last_login.mk` to remote sites
This file records the last login to the site where it is located, so
synchronizing it leads to wrong data on the remote sites.

Also, we have observered a CI failure where this file was edited (due to a user
login) while a snapshot for synchronization was created. This change should
mitigate this issue.

Change-Id: I4f9b1fe9d1ef9a0efc7c617ba362211e11c63bff