17817 FIX UpdateAction: remove invalid user profiles
In previous versions of Checkmk, when attempting to login with
an invalid userid, a profile directory was created. However,
this could cause problems when updating to 2.2, as these profile
directories could contain invalid characters.
This werk addresses this issues by first deleting any user
profile that isn't also a valid Checkmk user.

Co-authored-by: Logan Connolly <logan.connolly@checkmk.com>

SUP-22522

Change-Id: I6b52f811c9970cd932fca7a27f4745a2dd30a844