15358 Check_MK Agent: Show registered host name
In larger organisations a user may not always have access to the
cmk-update-agent.state file. This makes it hard to detect a mismatch
between the registered hostname and the actual hostname in checkmk.

Now the registered host name is shown along other agent information in
the "Check_MK Agent" service.

SUP-22634

Change-Id: I246a69bcabd3573e75c94bb315f5741bae113bea