17582 FIX Notifications: migrated parameters will include all rule numbers in their description
When upgrading to 2.4, notification rules and notification parameters will
be separated and saved to disk as different objects. See werk #17698.
In the description of the parameter objects, we included the rule number
like so.

"Migrated from notification rule #3"

However, it's possible that the same notification parameter was being
used in more than one notification rule. In this case, and with this werk,
we will now show all notification rule numbers in the parameter description.
For example.

"Migrated from notification rules #3, #7, #12"

Change-Id: I007e93fb1450d3c7b396d3f007525f76c6c4eb17