pylint checker: add allowed modules
The active hardware/software inventory check was moved
to a dedicated file. In order to put it where we ususally
put these kind of files, we need to allow this particular
plugin family to import from the cmk code.

This reflects the reality that these "plugins" depend
on aspects of checkmk that are not part of any official
plugin API.

We choose a more transparent approch here, by explicitly
adding the allowed imports.

Change-Id: Ic9228884681ca2b928e0edace95e1c721045a5b2