17714 FIX Fix change detection of clustered services for discovered changes
On clustered services the discovery parameters are written to each individual autochecks
file, but when the currently active node was not the first node in the cluster the
Service discovery detected this a a changed service, even though the correct parameters
have already been written to the correct nodes autocheck file.

Now, active nodes are being prioritized when detecting changed services which means this
problem will no longer occur.

CMK-22351

Change-Id: I5342a0e27a61646848995f2f032b4c47de610486