17529 FIX apc_ats_status: Handle undefined power supply status
When a "0" in the SNMP data indicates the absence of a power supply (e.g. because the unit does not support this) a `ValueError` got raised during parsing.

Now this value gets handled without a crash or warning. The absense of the power supply gets noted as "xV power supply not available".

SUP-22815

Change-Id: Ie65ac7668e56faa67902ce8ece054cbe126a5e05