printer_supply: factor out get supply color
This is groundwork for beginning to have a more feature-rich color
detector, e.g. support for multiple languages. Also, by introducing the
literal `Color` type we can constrain what colors are valid for printer
supply toners.

It's nearly a pure refactor. Howerever, we don't want to overlook raw
colors that are capitalized, e.g. "Black" should be matched as "black."
So we're now casting the raw color to lower before comparing in the
helper.

The only place that is affected downstream is the "color_info" logic,
but since it does a falsey check instead of checking for an empty
string, we're fine with having `None` as the default when no color is
present.

Change-Id: I2aac23f7bb64230cb4b6f452fe93433357414ec9