Cleanup orphaned discovered host label files on central site
Once every minute, a cron job syncs the discovered host label files from all
remote sites to the central site. They are needed in the central site by the
agent bakery. When a host is removed, the discovered host label file is removed
with it, but only in the site where the host was monitored. If the host was
monitored in a remote site, the file stayed behind in the central site.

As of this change, the sync job runs an additional cleanup routine which removes
such orphaned files. The cleanup is only executed once per hour because it
requires loading all known hosts, which is too expensive to be executed every
minute.

CMK-21782

Change-Id: I937b071b89f08c6f7e7a28ece1f9b9ad8f3d0e9a